requirements.txt
alembic==1.7.7
Authlib==0.15.5
Babel==2.10.1
bcrypt==3.2.0
bidict==0.22.0
blinker==1.4
boto3==1.20.54
botocore==1.23.54
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==8.1.2
cryptography==3.4.8
cycler==0.11.0
dnspython==2.2.1
email-validator==1.2.0
eventlet==0.33.0
Flask==2.1.1
Flask-Babel==2.0.0
Flask-Compress==1.11
Flask-Gravatar==0.5.0
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.4
Flask-SocketIO==5.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.33.2
gensim==4.2.0
greenlet==1.1.2
httpagentparser==1.9.2
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.4.2
ldap3==2.9.1
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
nose==1.3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.2
paramiko==2.10.4
passlib==1.7.4
Pillow==9.1.0
psutil==5.9.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
PyNaCl==1.5.0
pyotp==2.6.0
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
python-dateutil==2.8.2
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2021.3
qrcode==7.3.1
requests==2.25.1
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
smart-open==6.0.0
speaklater3==1.4
SQLAlchemy==1.4.35
sqlparse==0.4.2
sshtunnel==0.4.0
threadpoolctl==3.1.0
ua-parser==0.10.0
urllib3==1.26.9
user-agents==2.2.0
Werkzeug==2.0.3
WTForms==3.0.1
